payment_service/requirements.txt
fastapi
uvicorn
pydantic
squareup
cassandra-driver

